[LB-130] refactor: 누락된 디렉터리 구조 변경
